Change MAX 6071 to 6070 on v3.0 BoM by hand
Unintentional v2.3 to 3.0 change: the 6070 is what we need. We should check if the 6071 or the 6070 is included in the PCB definition, now or once we transition to KiCAD.

We noted this error because of faulty (doubled) battery-voltage readings in the cold.

From Bobby Schulz, who found the source of the error:

This is important because the MAX6071 does not have an internal filter and as a result the pinout changes - see Page 19 of datasheet. This pinout changes pin 1 from the filtered input (how we intended to use it) to 'Gound Force' - which has little documentation but looks to match the topology of the force/sense outputs but for low side switched applications. The upshot is that pin 1 goes from being a filter input connected to ground via a 0.1uF cap to the main ground connection. I think the cold is changing the impedance of the capacitor and causing the device to enter cutoff mode since the ground connection is now high impedance.

This issue can be resolved by switching the parts to the MAX6070, or by simply shorting pins 1 and 2 with a blob of solder. I have tested this fix and the device returns to the perfect stability we had been observing (1.8004V, and no deviation when placed in freezer)